fix(pkgs): moved `vim.nix` back into `nixos/` dir
`nixos-rebuild` doesn't recognize `(import ../vim.nix)`, possibly
because it thinks it's still working in `/etc/nixos/` even though I use
it from (and think of it as) `~/.config/nixos/`. Whatever the reason, I
won't be able to import `*.nix` files from higher directories in the
hierarchy, at least for the moment.