refactor(hardware): use EDID instead of DFP values
switch from DFP values (which can change based on which GPU port the
monitor is connected to) to persistent hardware-based EDID values (which
will always be the same for a given monitor). This may cause issues down
the line if I ever have a system with two monitors with the same EDID,
but that's unlikely and is a non-issue at present.