fix(X11): correctly set display size/resolutions
On startup, the X server wasn't detecting/setting the resolutions and
positions of each monitor correctly, resulting in a wonky looking SDDM
interface. This update sets `xrandr` variables correctly for a display
setup with a 4k 144hz monitor on the left and a 1440p 144hz monitor on
the right. The 1440p monitor is rotated clockwise (physically; the
display is rotated counter-clockwise to compensate). The monitors are
aligned on the center of their respective connecting edges. There are
still some issues with DPI/scaling, but this update marks a major
improvement.