rework: remove user `ash`
remove user `ash` and all related files and directories in order to
simplify user management for the time being; `ash` was an alternate
account used to separate "work" from "play" (`hunter` is/was "work,"
`ash` was "play"). Since this is not strictly necessary and Nix Flakes
and `home-manager` are hard enough to learn already, it makes sense to
consolidate `hunter` and `ash` into one account (at least for the
moment) and then only revisit the idea after successful configuration of
both Flakes and `home-manager`.