rework(`users`): temporary/untested flake rewrite
this commit will be overwritten/amended in the near future when I
continue the flake.nix user management rework project. As of now, the
changes in this commit have not been tested and are unlikely to work
properly. The only reason I'm committing these changes at all is because
I need to run `nixos-rebuild` with an unmodified `flake.nix` and I
didn't want to lose the progress I've made so far.